Refactor for some more clarity
